Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-core
langchain-google-genai
google-cloud-aiplatform
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart